restructured project based on current master. tested in vagrant, all routes working.
